changes to display
cleaned up some of the display, display is still not good but it is slightly better, needs a proper layout.

changed the display logic ot include "grouping" which removes all display results with identical paint matches. This feature needs to be better handled to show color groups in the source list.

select only three color swatches to display instead of four, the fourth was often wildly incorrect.